// types/EscrowHub.ts
export interface EscrowResponse {
  id: BigNumber;
  0: BigNumber;
  cid: string;
  1: string;
  buyer: string;
  2: string;
  seller: string;
  3: string;
  amount: BigNumber;
  4: BigNumber;
  fee: BigNumber;
  5: BigNumber;
  createdAt: BigNumber;
  6: BigNumber;
  expireAt: BigNumber;
  7: BigNumber;
  clearAt: BigNumber;
  8: BigNumber;
  state: number;
  9: number;
}
export interface DataResponse {
  id: BigNumber;
  0: BigNumber;
  cid: string;
  1: string;
  buyer: string;
  2: string;
  seller: string;
  3: string;
  amount: BigNumber;
  4: BigNumber;
  fee: BigNumber;
  5: BigNumber;
  createdAt: BigNumber;
  6: BigNumber;
  expireAt: BigNumber;
  7: BigNumber;
  clearAt: BigNumber;
  8: BigNumber;
  state: number;
  9: number;
}
export interface FetchEscrowsPaginatedResponse {
  data: DataResponse[];
  0: DataResponse[];
  totalItemCount: BigNumber;
  1: BigNumber;
  hasNextPage: boolean;
  2: boolean;
  nextCursor: BigNumber;
  3: BigNumber;
  length: 4;
}